MyFoodManage/requirements.txt
django
django-environ
django-widget-tweaks
django-bootstrap4
django-bootstrap-datepicker-plus